Implement REST api that can create and get tasks saved in the database
Axum is used for the web framework and sqlx is used for connecting to
the database.
The database is a sqlite database for simplicity.

The folder migrations contains sql commands that setup a sql database
from scratch.
It is also used to upgrade an older sql database schema to a newer one.
Sqlx keeps track of the schema version the database uses.

To build you need some one-time setup:
First create a file named .env in the root of the project.
This contains the connection string used for connecting to the database.
It can look like this:
```
DATABASE_URL=sqlite://dev.db
```
Then you need to install the sqlx commandline tool.
Use it to create the database and run the available migrations.
```
cargo install sqlx-cli
cargo sqlx db create
cargo sqlx migrate run
```