Add authentication
Add endpoints `/accounts/register` & `/accounts/login`

When you register or login successfully you receive an auth token.
This needs to be sent in the X-Auth-Token header in authenticated
endpoints.
To make an endpoint authenticated add the Authentication extractor to
the endpoint.
This will make the id of the account available to the endpoint.
If the authentication token is invalid the request terminates with an
error and the endpoint handler is never executed.